fix: 更新 FingerprintsBody 和 FingerprintResult 结构体中的 fingerprints 字段类型为 i64
[2025-06-20T02:12:52Z DEBUG actix_web::middleware::logger] Error in response: BadRequest("Json deserialize error: invalid value: integer `2413029736`, expected i32 at line 1 column 27")

+++ src/models/curseforge/requests.rs
@@ -76,7 +76,7 @@ pub struct FileIdsBody {
 
 #[derive(Serialize, Deserialize, Debug, ToSchema)]
 pub struct FingerprintsBody {
-    pub fingerprints: Vec<i32>
+    pub fingerprints: Vec<i64>
 }
 
 #[derive(Serialize, Deserialize, Debug, ToSchema)]
